Add edit election form and shared ElectionForm component
