+++ backend/src/Controllers/auth.controllers.ts
@@ -1,6 +1,7 @@
 var jwt = require('jsonwebtoken')
 
 const getUser = (req: any, res: any, next: any) => {
+  console.log("getUser")
   //This will not verify whether the signature is valid, need to add verification
   req.user = jwt.decode(req.cookies.id_token)
   next()
